backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyMySQL==1.1.2
